requirements.txt
# Code requirements

# Documentation
sphinx>=1.0
sphinx-rtd-theme

# Testing
pytest-cov>=2.0